cursor management for enter for split paragraph
