+++ src/lib/view/collection/section/default/control/CollapseIcon.svelte
