+++ src/lib/view/collection/section/default/control/ExpandIcon.svelte
