+++ src/lib/view/collection/section/default/control/SummaryIcon.svelte
